REFATORANDO

+++ Graph/Lib/Edge.cs
@@ -3,5 +3,5 @@ namespace Lib;
 public sealed record Edge(
     Node From,
     Node To,
-    int Weight
+    float Weight
 );
